Implement flexible toolbar resize
Due to limitations of ActionToolbar, it only relies on preferred sizes
of its components. To work around this, we must adjust these preferred
sizes to our liking. Fortunately, both our toolbars have only one
resizable component in it, which allows us to calculate the right size
by simply subtracting the sum of other component sizes and the toolbar
insets from the available size.

To determine the available size, we replace the top container that contains
both toolbars with our own component. This way we can take its width,
half it, and use this half as the area reserved for the toolbar.

Make sure to make the preferred width to be at least the minimum width,
in order for which to work we set isUsePreferredSizeAsMinimum for both
combo boxes, as otherwise they have the same minimum size as
the preferred size, which defeats the whole point of trying to make them
smaller if the preferred size doesn't fit.